requirements.txt
pandas
ddgs
pillow
requests
fastapi 
uvicorn